requirements.txt
Django==3.1.4
coverage==5.3.1
djangorestframework==3.12.2
